Fix alignment violation causing OUT_OF_RESOURCES error on Nvidia GPUs
I tested this fix on an Nvidia RTX 2070.

The issue stems from the fact that `reverse()` casts `lds` to `T2 *`,
whereas `lds` is defined in `tailFused` as `T []`, which is a weaker
alignment.

The workaround is to define the array as a `T2 []` of half size, which
will ensure correct alignment.