Use new bitlen GPU-side.
Integrates a proposal from George Woltman to speed-up the isBigWord()
GPU-side by avoiding the modulo N which is slow when N is NPOT.

Also removes using the sign of the weights as a flag for bigWord, doing
the computation instead. Tthis is a tiny slowdown right now, but allows
changing the weight representation in the future.