Best kcos implementation, ROCM bug worked around
Finally worked around Rocm bug with the best kcos implementation.  One extra useless if statement did the trick.  Nvidia users will get the unadulterated kcos code.
Eliminated first barrier in MiddleIn/MiddleOut.  The rocm optimizer was sometimes moving this so far forward in the instruction stream that memory loads did not have a chance to complete.  This is why the previous kcos implementation was faster despite generating a ton of extra ops.