Extract kernel tailSquareZero from tailSquare
This change extracts the special-case part of tailSquare into a
separate kernel. This simplified the remaining tailSquare.

The two kernels are still executed sequentially (which has a time penalty).
The slowdown is about 19us on Rpro7 single-worker and a bit less dual-worker.

Also allow kernel work-size to be specified as nb. of workgroups.